RABSW-192: Change fileshare PUT to allow mount/unmount (#85)
This commit changes the FileShare PUT code to allow the FileShare to change
after it has been created. Currently the only field that can be changed is
the FileSharePath. This allows the file share to be mounted and unmounted.

This required changes to separate out the mount/unmount code from the
create/delete file system code.

Co-authored-by: Matt Richerson <mattr@cray.com>